fix(updater): cast WAIT_TIMEOUT for windows wait status compare
windows.WaitForSingleObject returns uint32, but windows.WAIT_TIMEOUT is
typed (syscall.Errno), so the switch failed to compile on windows/amd64
and broke the release CLI build. Cast the constant to uint32 to match.

+++ internal/updater/process_windows.go
@@ -24,7 +24,7 @@ func isProcessRunning(pid int) (bool, error) {
 		return false, err
 	}
 	switch status {
-	case windows.WAIT_TIMEOUT:
+	case uint32(windows.WAIT_TIMEOUT):
 		return true, nil
 	case windows.WAIT_OBJECT_0:
 		return false, nil
